#7: Update command package references to reflect rename of cmd to command to avoid conflicts with built-in cmd module imported by some IDE debuggers
